Backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-bcrypt==1.0.1
PyJWT==2.9.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.2
statsmodels==0.14.2
numpy==1.26.4
